// Frontend/src/app/@core/auth/models/routes.enum.ts
export enum AppRoutes {
  Account = 'account',
  Users = 'users',
  Modules = 'modules',
  Assessments = 'assessments',
  CreateAssessment = 'create-assessment',
  EditAssessment = 'edit-assessment',
  Classes = 'classes',
  ClassMetrics = 'class-metrics',
  Class = 'class',
  Home = 'home',
  Settings = 'settings',
  Billing = 'billing',
  Logout = 'auth/logout',
  Students = 'students',
  ViewAssessment = 'view-assessment',
  AssessmentWritingDetails = 'assessment-writing-details',
  ClassAnalysis = 'class-analysis',
}